adjust namespaces and move scripts to folders
