--- Assets/Scripts/BurgerView.cs
@@ -1,3 +1,0 @@
-﻿public class BurgerView : CookableItemView
-{
-}
