--- Assets/Scripts/EventStreams.cs
@@ -1,7 +1,0 @@
-using SimpleEventBus;
-using SimpleEventBus.Interfaces;
-
-public static class EventStreams
-{
-    public static IEventBus Game { get; } = new EventBus();
-}
