+++ Assets/Scripts/Events/EventStreams.cs
@@ -1,0 +1,10 @@
+using SimpleEventBus;
+using SimpleEventBus.Interfaces;
+
+namespace Events
+{
+    public static class EventStreams
+    {
+        public static IEventBus Game { get; } = new EventBus();
+    }
+}
