+++ Assets/Scripts/Events/ReleaseMeatRequestEvent.cs
@@ -1,0 +1,15 @@
+using Controllers;
+using SimpleEventBus.Events;
+
+namespace Events
+{
+    public class ReleaseMeatRequestEvent : EventBase
+    {
+        public MeatController MeatController { get; }
+
+        public ReleaseMeatRequestEvent(MeatController meatController)
+        {
+            MeatController = meatController;
+        }
+    }
+}
