+++ Assets/Scripts/Models/PanModel.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using UnityEngine;
+
+namespace Models
+{
+    [Serializable]
+    public class PanModel
+    {
+        [field: SerializeField]
+        public bool IsAvailable { get; set; }
+    
+        [field: SerializeField]
+        public float CookingTime { get; private set; }
+
+        public PanModel(bool isAvailable, float cookingTime)
+        {
+            IsAvailable = isAvailable;
+            CookingTime = cookingTime;
+        }
+    }
+}
