+++ Assets/Scripts/Models/SaveDataModel.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Models
+{
+    [Serializable]
+    public class SaveDataModel
+    {
+    
+    }
+}
