--- Assets/Scripts/PanController.cs
@@ -1,55 +1,0 @@
-﻿using System;
-using UnityEngine;
-using UnityEngine.EventSystems;
-
-public class PanController : MonoBehaviour, IDropHandler
-{
-    [SerializeField]
-    private PanView _panView;
-    
-    private PanModel _panModel;
-    private MeatController _meatController;
-    private IDisposable _subscription;
-    
-    public void Initialize(PanModel panModel, MeatController meatController)
-    {
-        _panModel = panModel;
-        _meatController = meatController;
-        _panView.Initialize(_meatController.MeatView);
-        _panModel.IsAvailable = false;
-        
-        _panView.MeatCooked += OnMeetCooked;
-        _subscription = EventStreams.Game.Subscribe<ReleaseMeatRequest>(Action);
-    }
-
-    private void Action(ReleaseMeatRequest obj)
-    {
-        _panModel.IsAvailable = true;
-    }
-
-    private void OnMeetCooked()
-    {
-        _meatController.ChangeMeatState(true);
-    }
-
-    private void Update()
-    {
-        if ( _panModel==null || _panModel.IsAvailable)
-        {
-            return;
-        }
-        
-        _panView.Cook(_panModel.CookingTime);
-    }
-
-
-    public void OnDrop(PointerEventData eventData)
-    {
-        Debug.Log("Сковородкин");
-    }
-
-    private void OnDestroy()
-    {
-        _subscription.Dispose();
-    }
-}
