--- Assets/Scripts/PlateModel.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using UnityEngine;
-
-[Serializable]
-public class PlateModel
-{
-    [field: SerializeField]
-    public bool IsAvailable { get; set; }
-
-    public PlateModel(bool isAvailable)
-    {
-       IsAvailable = isAvailable;
-    }
-    
-}
