--- Assets/Scripts/ReleaseMeatRequest.cs
@@ -1,11 +1,0 @@
-using SimpleEventBus.Events;
-
-public class ReleaseMeatRequest : EventBase
-{
-    public MeatController MeatController { get; }
-
-    public ReleaseMeatRequest(MeatController meatController)
-    {
-        MeatController = meatController;
-    }
-}
