--- Assets/Scripts/SaveDataModel.cs
@@ -1,7 +1,0 @@
-using System;
-
-[Serializable]
-public class SaveDataModel
-{
-    
-}
