+++ Assets/Scripts/Services/IFileService.cs
@@ -1,3 +1,5 @@
+using Models;
+
 namespace Services
 {
     public interface IFileService : IService
