+++ Assets/Scripts/Views/BurgerView.cs
@@ -1,0 +1,6 @@
+﻿namespace Views
+{
+    public class BurgerView : CookableItemView
+    {
+    }
+}
