+++ Assets/Scripts/Views/MeatView.cs
@@ -1,5 +1,8 @@
-﻿public class MeatView : CookableItemView
+﻿namespace Views
 {
+    public class MeatView : CookableItemView
+    {
+    }
 }
